+++ example/MultiCast Communication/Emitter2/Emitter2.ino
@@ -3,7 +3,7 @@
 #include "Mirf.h"
 
 Nrf24l Mirf = Nrf24l(10, 9); // CE,CSN
-int value = 200;
+int16_t value = 200;
 
 void setup()
 {
